src/app/requirements.txt
Flask
psycopg2
python-dotenv
configparser
